python/requirements.txt
azure-functions
azure-devops
requests